Add `policy-pack` configuration option
This allows us to run a local policy pack whenever we run Pulumi,
using the `--policy-pack` option. This will be in addition to any
global policies that may be active for the stack in the Pulumi SaaS.

This will allow us to more easily begin experimenting with policy for
our infrastructure.

Tests are added, including the use of a local policy pack in our
Buildkite pipeline. Pretty meta, right?

Signed-off-by: Christopher Maier <chris@graplsecurity.com>