Update `buildkite/plugin-tester` image to v2.0.0
There's finally a formal release of `buildkite/plugin-tester` that we
can use! This locks us to that new release, and addresses a breaking
change that was introduced.

See
https://github.com/buildkite-plugins/buildkite-plugin-tester/releases/tag/v2.0.0

(We also update `coreutils` in our "upgrade" of the `plugin-tester`
image to the one that is current in the update Alpine container we're
now using.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>